Rework conversion tests to allow for stress testing and memory leaks
Because Crystal and JSC have their own garbage collectors, things can vanish from one or the other side unless references are propperly retained.
The multiple tests here insure that things dont' vanish when they shouldn't.
Before the extra Context#protect method was added in the last commit, the Context would vanish while Crystal was still referencing it.
To complicate matters, the JSC source isn't _perfectly _explicit on when a reference count is increased.
(Search for "Create Rule" in the API source to see when it _is specified.)